Fixed some issues and Animated Bg
I add some BackGround animations, that makes the game more alive. Added the following: Animated Bg Nightclub , animated Bg Airport, and replaced the Port BG to one of a Cascade in jungle. Also that one is animated. The only one that still a picture is the Forest (Working on him). 

Added with all that stuff mentioned, i begun to animate the other Boy Champ (John), where there is only one anim, the IDLE one.