Added the direction
With the direction the players turn to face the other one. But for now the textures only do that. We need to flip also the Box Colliders for each player, making that with the help of the vector Direction.

Also the next to make with the boxes are the HelathBars for each player.